Enable multiple activity spotting
This changes Spotter Control from POTA to a standard control, but
leaving and introducting a spot posting function as part of the
extension.

This enables now for a single spotting control that can spot multiple
activities. The activities to be spotted are displayed to user via Badge
items on spot button.

This PR also includes WWFF spotting as a functional example to show
multiple spotting.
